+++ Changelog.md
@@ -8,6 +8,7 @@ WhateverGreen Changelog
 - Added `framebuffer-camellia` and `framebuffer-flags` patching
 - Added `AppleBacklight` patching code (based on `AppleBacklightFixup` by `hieplpvip`)
 - Enabled CFL backlight patches by default on laptops with CFL graphics
+- Added NVIDIA HDMI enabling code for firmwares that disable it by default (thanks @Fraxul)
 
 #### v1.2.4
 - Added platform list dumping to ioreg (at IOService:/IOResources/WhateverGreen), debug build only with -igfxfbdump